requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
psycopg2-binary
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorlog==6.6.0
frozenlist==1.3.0
hikari
hikari-lightbulb==2.1.3
hikari-miru==0.6.3
idna==3.3
lavasnek-rs
multidict==5.2.0
mutagen==1.45.1
pure25519==0.0.1
pycryptodomex==3.14.0
requests==2.27.1
songbird-py==0.1.5
typing_extensions==4.0.1
urllib3==1.26.8
uvloop==0.16.0
websockets==10.1
yarl==1.7.2
youtube-search==2.1.0
yt-dlp==2021.12.1
